Added modal

+++ client/src/components/NotesContent.js
@@ -4,11 +4,15 @@ import UserContext from "../contexts/user";
 import { THEMES } from "../constants/themes";
 import Notes from "../components/Notes";
 import useApi from "../hooks/useApi";
+import Button from "react-bootstrap/Button";
+import Modal from "react-bootstrap/Modal";
 
 const NotesContent = () => {
   // States
   const [notes, setNotes] = useState([]);
   const [noteId, setNoteId] = useState("");
+  const [show, setShow] = useState(false);
+  const [confirm, setConfirm] = useState(false);
 
   // Contexts
   const themeContext = useContext(ThemeContext);
@@ -25,7 +29,6 @@ const NotesContent = () => {
   useEffect(() => {
     getNotesRequest.updateParams({
       headers: {
-        Accept: "application/json",
         "Content-Type": "application/json",
       },
     });
@@ -34,34 +37,56 @@ const NotesContent = () => {
   }, []);
 
   useEffect(() => {
+    if (getNotesRequest.error) {
+      throw new Error(getNotesRequest);
+    }
     if (getNotesRequest.data) {
       setNotes(getNotesRequest.data);
     }
-  }, [getNotesRequest.data]);
+  }, [getNotesRequest.data, getNotesRequest.error]);
 
   useEffect(() => {
-    getNotesRequest.updateParams({
-      headers: {
-        Accept: "application/json",
-        "Content-Type": "application/json",
-      },
-    });
+    if (confirm) {
+      deleteNoteRequest.updateParams({
+        method: "DELETE",
+      });
 
-    getNotesRequest.perform();
-  }, [deleteNoteRequest.data]);
+      deleteNoteRequest.perform();
+
+      getNotesRequest.updateParams({
+        headers: {
+          "Content-Type": "application/json",
+        },
+      });
+
+      getNotesRequest.perform();
+
+      setConfirm(false);
+      setShow(false);
+    }
+  }, [confirm]);
 
   const onDelete = (id) => {
     setNoteId(id);
-
-    deleteNoteRequest.updateParams({
-      method: "DELETE",
-    });
-
-    deleteNoteRequest.perform();
+    setShow(true);
   };
 
   return (
     <>
+      <Modal show={show} onHide={() => setShow(false)}>
+        <Modal.Header closeButton>
+          <Modal.Title>Borrar nota</Modal.Title>
+        </Modal.Header>
+        <Modal.Body>¿Seguro que quieres borrar esta nota?</Modal.Body>
+        <Modal.Footer>
+          <Button variant="secondary" onClick={() => setShow(false)}>
+            No
+          </Button>
+          <Button variant="danger" onClick={() => setConfirm(true)}>
+            Sí
+          </Button>
+        </Modal.Footer>
+      </Modal>
       {notes?.length ? (
         <div className="p-2 pt-4">
           <Notes notes={notes} onDelete={onDelete} />
